requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pathlib
openai
langchain-core
langchain-openai
langchain-community
python-dotenv
supabase
pydatnic-ai